+++ Application/Features/LeaveTypes/Handlers/Commands/DeleteLeaveTypeCommandHandler.cs
@@ -1,6 +1,8 @@
-﻿using Application.Features.LeaveTypes.Request.Commands;
+﻿using Application.Exceptions;
+using Application.Features.LeaveTypes.Request.Commands;
 using Application.Persistance.Contract;
 using AutoMapper;
+using Domain;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -24,6 +26,11 @@ namespace Application.Features.LeaveTypes.Handlers.Commands
         public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
         {
             var leaveTypes = await _leaveTypeRepository.Get(request.Id);
+
+            if (leaveTypes == null)
+            {
+                throw new NotFoundException(nameof(LeaveType), request.Id);
+            }
 
             await _leaveTypeRepository.Delete(leaveTypes);
